+++ python/lsst/ts/wep/utils/__init__.py
@@ -2,7 +2,7 @@ from .enumUtils import *
 from .ioUtils import *
 from .maskUtils import *
 from .miscUtils import *
+from .modelUtils import *
 from .plotUtils import *
 from .taskUtils import *
-from .testUtils import *
 from .zernikeUtils import *